fix: use importlib_metadata on python <3.10
We updated our version of importlib_metadata, so we might as well take
advantage of it to have full compatibility with the features introduced
in Python 3.10.

If this change is not desired, we can easily drop it. If it's
uncontroversial, we can just squash it down when the Cleo PR is merged
and this PR is rebased to depend on a Cleo release.